requirements.txt
Requests>=2.31.0
wom.py>=0.6.1